+++ CHANGES.md
@@ -3,6 +3,7 @@ merlin NEXT_VERSION
 
   + merlin binary
     - Fix a follow-up issue to the preference of non-ghost nodes introduced in #1660 (#1690, fixes #1689)
+    - Add `--cache-period` flag, that sets cache invalidation period. (#1698)
   + editor modes
     - vim: load merlin when Vim is compiled with +python3/dyn (e.g. MacVim)